+++ src/singletons/Settings.hpp
@@ -345,6 +345,13 @@ public:
     QStringSetting preferredQuality = {"/external/streamlink/quality",
                                        "Choose"};
     QStringSetting streamlinkOpts = {"/external/streamlink/options", ""};
+    BoolSetting streamlinkOptsLatency = {"/external/streamlink/optionsLatency",
+                                         true};
+    BoolSetting streamlinkOptsAds = {"/external/streamlink/optionsAds", true};
+
+    // VLC attached player
+    BoolSetting vlcFollowActive = {"/external/vlc/followActive", true};
+    QStringSetting vlcPlayerPath = {"/external/vlc/path", ""};
 
     // Custom URI Scheme
     QStringSetting customURIScheme = {"/external/urischeme"};