+++ src/widgets/dialogs/QualityPopup.hpp
@@ -11,8 +11,12 @@ namespace chatterino {
 class QualityPopup : public BasePopup
 {
 public:
-    QualityPopup(const QString &channelURL, QStringList options);
-    static void showDialog(const QString &channelURL, QStringList options);
+    QualityPopup(const QString &channelURL, QStringList options,
+                 QStringList extraArguments = QStringList(),
+                 bool streamVLC = false);
+    static void showDialog(const QString &channelURL, QStringList options,
+                           QStringList extraArguments = QStringList(),
+                           bool streamVLC = false);
 
 protected:
     void keyPressEvent(QKeyEvent *e) override;
@@ -28,6 +32,8 @@ private:
     } ui_;
 
     QString channelURL_;
+    QStringList extraArguments_;
+    bool streamVLC_;
 };
 
 }  // namespace chatterino